fix - code optimizations
